Revert "making the markodown more modular"
This reverts commit 5ec4706bc821e20ed7e066da4d231fa0a15fb976.
