--- components/markdown/index.ts
@@ -1,17 +1,0 @@
-export {
-  parseMarkdown,
-  transformMarkdownLinks,
-  type Element,
-} from './MarkdownParser';
-export { MarkdownRenderer } from './MarkdownRenderer';
-export { TabsGroup } from './TabsGroup';
-export { TableOfContents, TableOfContentsWrapper } from './TableOfContents';
-
-// Export overrides for potential customization
-export { textOverrides } from './overrides/TextElements';
-export { listOverrides } from './overrides/ListElements';
-export { tableOverrides } from './overrides/TableElements';
-export { codeOverrides } from './overrides/CodeElements';
-export { infoBoxOverrides } from './overrides/InfoBoxes';
-export { customOverrides } from './overrides/CustomElements';
-export { headlineOverrides } from './overrides/HeadlineElements';
