--- components/markdown/overrides/HeadlineElements.tsx
@@ -1,13 +1,0 @@
-import {
-  Headline1,
-  Headline2,
-  Headline3,
-  Headline4,
-} from '~/components/Headlines';
-
-export const headlineOverrides = {
-  h1: { component: Headline1 },
-  h2: { component: Headline2 },
-  h3: { component: Headline3 },
-  h4: { component: Headline4 },
-};
